backend/requirements.txt
fastapi==0.110.0
pydantic==2.6.3
uvicorn==0.28.0
ipython==8.21.0
pgcli==4.0.1
psycopg-binary==3.1.13
psycopg-pool==3.2.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.28
alembic==1.13.1
python-dotenv==1.0.1
python-multipart==0.0.9
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]